Removed user profile related code and added signaling mechanism
